refactor: saparate code that fetches content and controlls ipfs node
