+++ config.json
@@ -1,0 +1,3 @@
+{
+  "ipfsRoute": "ipfs"
+}
