+++ src/config.js
@@ -1,0 +1,6 @@
+import preval from 'preval.macro';
+
+export default preval`
+  const fs = require('fs');
+  module.exports = JSON.parse(fs.readFileSync('./config.json'));
+`;
