+++ src/ipfs-service-worker/ipfsNode.js
